feat(interfaces): Introduce programmatic editor for UI interfaces
This introduces `InterfacesCSTEditor` to enable programmatic parsing
and generation of UI interface definitions. Leveraging LibCST, this
new editor allows for automated management of the `Interfaces` class
members and the `_gen_interfaces` method. This reduces manual errors,
ensures consistency, and streamlines future updates to the game's
UI interfaces by providing a structured, code-driven approach.
It removes the wildcard import from `.navigator` as this file now
serves as a tool for interface definition, not direct runtime usage.

[skip ci]